backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-genai>=0.4.0
firecrawl-py>=1.0.0
supabase>=2.0.0
boto3>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
Pillow>=10.0.0
